requirements.txt
absl-py==2.1.0
astunparse==1.6.3
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
Flask==3.1.0
Flask-Cors==5.0.0
flatbuffers==24.3.25
fonttools==4.55.0
gast==0.4.0
google-auth==2.36.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.68.1
gunicorn==23.0.0
h5py==3.12.1
idna==3.10
itsdangerous==2.2.0
jax==0.4.30
jaxlib==0.4.30
Jinja2==3.1.4
keras==2.12.0
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.9.3
ml_dtypes==0.5.0
numpy==1.23.5
oauthlib==3.2.2
opt_einsum==3.4.0
packaging==24.2
pillow==11.0.0
protobuf==4.25.5
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
scipy==1.14.1
six==1.16.0
tensorboard==2.12.3
tensorboard-data-server==0.7.2
termcolor==2.5.0
typeguard==2.13.3
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
wrapt==1.14.1
